requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
beautifulsoup4==4.9.1
camelot-py==0.8.2
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
ci-info==0.2.0
click==7.1.2
colorama==0.4.3
comtypes==1.1.7
configobj==5.0.6
configparser==5.0.1
console==0.9905
cryptography==3.2.1
decorator==4.4.2
distlib==0.3.1
distro==1.5.0
Django==3.1.2
docopt==0.6.2
Elixir==0.7.1
et-xmlfile==1.0.1
etelemetry==0.2.2
ezenv==0.92
filelock==3.0.12
fire==0.3.1
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
fpdf==1.7.2
future==0.18.2
ghostscript==0.6
gTTS-token==1.1.3
httplib2==0.18.1
idna==2.10
img2pdf==0.4.0
isodate==0.6.0
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
Js2Py==0.70
lazy-object-proxy==1.4.3
lxml==4.6.1
MarkupSafe==1.1.1
mccabe==0.6.1
mutagen==1.45.1
mysql-connector-python==8.0.21
networkx==2.5
nibabel==3.2.0
nipype==1.6.0
numpy==1.19.1
opencv-python==4.4.0.46
openpyxl==3.0.5
packaging==20.4
pandas==1.1.4
pathlib==1.0.1
pbr==5.5.1
pdf2image==1.14.0
pdfminer.six==20201018
pikepdf==2.0.0
Pillow==8.1.0
pipwin==0.5.0
protobuf==3.13.0
prov==2.0.0
PyAudio @ file:///C:/Users/aakan/pipwin/PyAudio-0.2.11-cp38-cp38-win_amd64.whl
pycparser==2.20
pydot==1.4.1
pydotplus==2.0.2
pygame==1.9.6
pyjsparser==2.7.1
pylint==2.5.3
PyMuPDF==1.18.5
PyMySQL==0.10.1
pyparsing==2.4.7
pypiwin32==223
PyPrind==2.11.2
pySmartDL==1.3.3
pytesseract==0.3.6
python-dateutil==2.8.1
python-docx==0.8.10
pyttsx3==2.90
pytube==9.7.0
pytube3==9.6.4
pytz==2020.1
pywin32==227
pyxnat==1.4
rdflib==5.0.0
requests==2.24.0
scipy==1.5.4
simplejson==3.17.2
six==1.15.0
sortedcontainers==2.3.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
SQLAlchemy==0.7.10
sqlalchemy-migrate==0.11.0
sqlparse==0.4.1
tabula-py==2.2.0
Tempita==0.5.2
termcolor==1.1.0
tkfontchooser==2.0.2
toml==0.10.1
tqdm==4.52.0
traits==6.1.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.10
virtualenv==20.0.35
Werkzeug==1.0.1
wikipedia==1.4.0
wrapt==1.12.1
xlrd==0.7.1
xlwt==0.7.2
youtube-dl==2020.11.1.1
